Added powershell module DCOMrade
The python module helper for the DCOMrade script I created.

DCOMrade can enumerate the possible vulnerable DCOM applications based on a set of values. A research paper will be released soon explaining the research and how it can be used. Outputs a list with interesting CLSID's (based on the amount of MemberTypes) and a list with DCOM applications that have MemberTypes which are present in the string subset provided in the script (can be changed to fit the needs of the user)