Added Invoke-DCOMrade.ps1
This is a powershell script that can enumerate possible vulnerable DCOM applications. The research is based on enigma0x3's findings but implemented in a script. The script outputs the potentially interesting CLSID's based on the amount of MemberTypes and outputs a list of vulnerable DCOM applications based on a list of strings that were derived from the other posts about abusing DCOM applications